+++ llama_index/embeddings/huggingface.py
@@ -1,13 +1,20 @@
-from typing import Any, List, Optional
+import asyncio
+from typing import Any, List, Optional, Sequence
 
 from llama_index.bridge.pydantic import Field, PrivateAttr
 from llama_index.callbacks import CallbackManager
-from llama_index.embeddings.base import DEFAULT_EMBED_BATCH_SIZE, BaseEmbedding
+from llama_index.embeddings.base import (
+    DEFAULT_EMBED_BATCH_SIZE,
+    BaseEmbedding,
+    Embedding,
+)
 from llama_index.embeddings.huggingface_utils import (
     DEFAULT_HUGGINGFACE_EMBEDDING_MODEL,
     format_query,
     format_text,
 )
+from llama_index.embeddings.pooling import Pooling
+from llama_index.llms.huggingface import HuggingFaceInferenceAPI
 from llama_index.utils import get_cache_dir, infer_torch_device
 
 
@@ -176,3 +183,111 @@ class HuggingFaceEmbedding(BaseEmbedding):
             format_text(text, self.model_name, self.text_instruction) for text in texts
         ]
         return self._embed(texts)
+
+
+class HuggingFaceInferenceAPIEmbeddings(HuggingFaceInferenceAPI, BaseEmbedding):  # type: ignore[misc]
+    """
+    Wrapper on the Hugging Face's Inference API for embeddings.
+
+    Overview of the design:
+    - Uses the feature extraction task: https://huggingface.co/tasks/feature-extraction
+    """
+
+    pooling: Optional[Pooling] = Field(
+        default=Pooling.CLS,
+        description=(
+            "Optional pooling technique to use with embeddings capability, if"
+            " the model's raw output needs pooling."
+        ),
+    )
+    query_instruction: Optional[str] = Field(
+        default=None,
+        description=(
+            "Instruction to prepend during query embedding."
+            " Use of None means infer the instruction based on the model."
+            " Use of empty string will defeat instruction prepending entirely."
+        ),
+    )
+    text_instruction: Optional[str] = Field(
+        default=None,
+        description=(
+            "Instruction to prepend during text embedding."
+            " Use of None means infer the instruction based on the model."
+            " Use of empty string will defeat instruction prepending entirely."
+        ),
+    )
+
+    @classmethod
+    def class_name(cls) -> str:
+        return "HuggingFaceInferenceAPIEmbeddings"
+
+    async def _async_embed_single(self, text: str) -> Embedding:
+        embedding = (await self._async_client.feature_extraction(text)).squeeze(axis=0)
+        if len(embedding.shape) == 1:  # Some models pool internally
+            return list(embedding)
+        try:
+            return list(self.pooling(embedding))  # type: ignore[misc]
+        except TypeError as exc:
+            raise ValueError(
+                f"Pooling is required for {self.model_name} because it returned"
+                " a > 1-D value, please specify pooling as not None."
+            ) from exc
+
+    async def _async_embed_bulk(self, texts: Sequence[str]) -> List[Embedding]:
+        """
+        Embed a sequence of text, in parallel and asynchronously.
+
+        NOTE: this uses an externally created asyncio event loop.
+        """
+        tasks = [self._async_embed_single(text) for text in texts]
+        return await asyncio.gather(*tasks)
+
+    def _get_query_embedding(self, query: str) -> Embedding:
+        """
+        Embed the input query synchronously.
+
+        NOTE: a new asyncio event loop is created internally for this.
+        """
+        return asyncio.run(self._aget_query_embedding(query))
+
+    def _get_text_embedding(self, text: str) -> Embedding:
+        """
+        Embed the text query synchronously.
+
+        NOTE: a new asyncio event loop is created internally for this.
+        """
+        return asyncio.run(self._aget_text_embedding(text))
+
+    def _get_text_embeddings(self, texts: List[str]) -> List[Embedding]:
+        """
+        Embed the input sequence of text synchronously and in parallel.
+
+        NOTE: a new asyncio event loop is created internally for this.
+        """
+        loop = asyncio.new_event_loop()
+        try:
+            tasks = [
+                loop.create_task(self._aget_text_embedding(text)) for text in texts
+            ]
+            loop.run_until_complete(asyncio.wait(tasks))
+        finally:
+            loop.close()
+        return [task.result() for task in tasks]
+
+    async def _aget_query_embedding(self, query: str) -> Embedding:
+        return await self._async_embed_single(
+            text=format_query(query, self.model_name, self.query_instruction)
+        )
+
+    async def _aget_text_embedding(self, text: str) -> Embedding:
+        return await self._async_embed_single(
+            text=format_text(text, self.model_name, self.text_instruction)
+        )
+
+    async def _aget_text_embeddings(self, texts: List[str]) -> List[Embedding]:
+        return await self._async_embed_bulk(
+            texts=[
+                format_text(text, self.model_name, self.text_instruction)
+                for text in texts
+            ]
+        )