+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Bug Fixes / Nits
 
+- Fixed additional kwargs in ReActAgent.from_tools() (#8206)
 - Fixed missing spaces in prompt templates (#8190)
 - Remove auto-download of llama2-13B on exception (#8225)