+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - add semantic kernel document format (#8226)
 - Improve MyScale Hybrid Search and Add `DELETE` for MyScale vector store (#8159)
 
+### New Features
+
+- Support for Hugging Face Inference API's `conversational`, `text_generation`,
+  and `feature_extraction` endpoints via `huggingface_hub[inference]` (#8098)
+
 ### Bug Fixes / Nits
 
 - Fixed additional kwargs in ReActAgent.from_tools() (#8206)